màj du projet

+++ app/Http/Controllers/EleveController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use App\Models\Eleve;
 use App\Models\ParentM;
+use Illuminate\Http\Response;
 
 class EleveController extends Controller
 {
@@ -12,17 +14,18 @@ class EleveController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return Response
+     * @return JsonResponse
      */
     public function index()
     {
-        return Eleve::select('Matricule',
-        'Nom','Prénom','الاسم العائلي','الاسم الشخصي','Sexe','Email', 'date_naissance','Adress','image')->get();
-    }
+        $eleves = Eleve::all();
+
+        return response()->json(['eleves' => $eleves], 200);
+    } 
     /**
      * Show the form for creating a new resource.
      *
-     * @return Response
+     * @return JsonResponse
      */
 public function create()
 {
@@ -30,65 +33,25 @@ public function create()
         'Nom','Prénom','الاسم العائلي','الاسم الشخصي','Sexe','Email', 'date_naissance','Adress','image')->post();
    }
 
-public function store(Request $request)
+// app/Http/Controllers/EleveController.php
+
+
+    public function store(Request $request)
 {
     $request->validate([
-        'Matricule' => 'required',
         'Nom' => 'required',
         'Prénom' => 'required',
-        'الاسم العائلي' => 'required',
-        'الاسم الشخصي' => 'required',
-        'Sexe' => 'required',
         'date_naissance' => 'required|date',
-        'Email' => 'required|Email|unique:eleves,Email',
-        'Adress' => 'required',
-        'image' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-        'parent_id' => 'required',
-        'type' => 'required'
-
     ]);
 
-    if ($request->hasFile('image')) {
-        $image = $request->file('image');
-        $name = time().'.'.$image->getClientOriginalExtension();
-        $destinationPath = public_path('/images');
-        $image->move($destinationPath, $name);
-        $imageUrl = url('/images/'.$name);
-    } else {
-        $imageUrl = '';
-    }
-
-    $eleve = new Eleve([
-        'Matricule'=>$request->get('Matricule'),
-        'Nom' => $request->get('Nom'),
-        'Prénom' => $request->get('Prénom'), 
-        'الاسم العائلي' => $request->get('الاسم العائلي'),
-        'الاسم الشخصي'=> $request->get('الاسم الشخصي'),
-        'Sexe' => $request->get('Sexe'),
-        'Email' => $request->get('Email'),
-        'Adress' => $request->get('Adress'),
-        'date_naissance' => $request->get('date_naissance'),
-        'Tél' => $request->get('Tél'),
-        'image' => $imageUrl,
-        'parent_id' => $request->get('parent_id'),
-        'type' => $request->get('type'), //  cette ligne est pour enregistrer le type de relation
-
-    ]);
-        $relation = [
-            'eleve_id' => $eleve->id,
-            'parent_id' => $request->get('parent_id'),
-            'type' => $request->get('type')
-        ];
-
-        $eleve = new eleve($relation);
-        
-
+    $eleve = new Eleve;
+    $eleve->Nom = $request->input('Nom');
+    $eleve->Prénom = $request->input('Prénom');
+    $eleve->date_naissance = $request->input('date_naissance');
     $eleve->save();
 
-    return response()->json([
-                'message'=>'Elève ajouté avec succès'
-            ]);
-        }
+    return response()->json($eleve, 201);
+}
 
 public function edit($id)
 {
@@ -125,8 +88,8 @@ public function update(Request $request, $id)
         $imageUrl = $eleve->image;
     }
 
-    $eleve->Nom = $request->get('nom');
-    $eleve->Prénom = $request->get('prenom');
+    $eleve->Nom = $request->get('Nom');
+    $eleve->Prénom = $request->get('Prénom');
     $eleve->date_naissance = $request->get('date_naissance');
     $eleve->Adress = $request->get('Adress');
     $eleve ->Matricule = $request->get('Matricule');
@@ -138,17 +101,22 @@ public function update(Request $request, $id)
 
 }
     public function show($id)
-    {
-        $eleve = Eleve::with('parent')->findOrFail($id);
-        return view('eleves.show', compact('eleve'));
-    }
+{
+    $eleve = Eleve::find($id);
+    return response()->json($eleve);
+}
+
 
     public function destroy($id)
-    {
-        $eleve = Eleve::findOrFail($id);
-        $eleve->delete();
+{
+    $eleve = Eleve::find($id);
 
-        return redirect()->route('eleves.index')
-        ->with('success', 'Eleve supprimé avec succès.');
+    if (!$eleve) {
+        return response()->json(['message' => 'L\'élève n\'existe pas'], 404);
     }
+                            
+    $eleve->delete();
+    return response()->json(['message' => 'L\'élève a été supprimé avec succès']);
+}
+
 }
