+++ app/Http/Controllers/ParentController.php
@@ -1,26 +1,39 @@
 <?php
 
 namespace App\Http\Controllers;
-
-use App\Models\ParentM;
 use Illuminate\Http\Request;
+use Illuminate\Http\JsonResponse;
+use App\Models\ParentM;
+use App\Models\Eleve;
+
 
 class ParentController extends Controller
 {
-    public function index()
+     /**
+     * Display a listing of the resource.
+     *
+     * @return JsonResponse
+     */
+   public function index()
     {
         $parents = ParentM::all();
-        return view('parents.index', compact('parents'));
-    }
 
+        return response()->json(['parents' => $parents], 200);
+    } 
+ /**
+     * Show the form for creating a new resource.
+     *
+     * @return JsonResponse
+     */
     public function create()
-    {
-        return view('parents.create');
-    }
+{
+    return ParentM::select('Civilité',
+        'Nom','Prénom','Tél','Identifiant')->post();
+   }
 
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
+        $request->validate([
             'Civilité' => 'required',
             'Nom' => 'required',
             'Prénom' => 'required',
@@ -30,25 +43,29 @@ class ParentController extends Controller
             'Identifiant' => 'required'
         ]);
 
-        $parent = new ParentM();
-        $parent->Civilité = $validatedData['Civilité'];
-        $parent->Nom = $validatedData['Nom'];
-        $parent->Prénom = $validatedData['Prénom'];
-        $parent->Email = $validatedData['Email'];
-        $parent->Adress = $validatedData['Adress'];
-        $parent->tél = $validatedData['tél'];
-        $parent->Identifiant = $validatedData['Identifiant'];
-
+        $parent = new ParentM([
+            'Identifiant' => $request->get('Identifiant'),
+            'Nom' => $request->get('Nom'),
+            'Prénom' => $request->get('Prénom'),
+            'Civilité' => $request->get('Civilité'),
+            'Email' => $request->get('Email'),
+            'Adress' => $request->get('Adress'),
+            'tél' => $request->get('tél'),
+        ]);
+       
         $parent->save();
 
-        return redirect()->route('parents.index')->with('success', 'Le parent a été ajouté avec succès.');
-    }
+        return response()->json([
+                'message'=>'Parent ajouté avec succès'
+            ]);   
+         }
 
-    public function show($id)
-    {
-        $parent = ParentM::findOrFail($id);
-        return view('parents.show', compact('parent'));
-    }
+        public function show($id)
+{
+    $parent = ParentM::find($id);
+    return response()->json($parent);
+}
+
 
     public function edit($id)
     {
@@ -58,7 +75,7 @@ class ParentController extends Controller
 
     public function update(Request $request, $id)
     {
-        $validatedData = $request->validate([
+        $request->validate([
             'Civilité' => 'required',
             'Nom' => 'required',
             'Prénom' => 'required',
@@ -68,24 +85,29 @@ class ParentController extends Controller
             'Identifiant' => 'required',
         ]);
 
-        $parent = ParentM::findOrFail($id);
-        $parent->Civilité = $validatedData['Civilité'];
-        $parent->Nom = $validatedData['Nom'];
-        $parent->Prénom = $validatedData['Prénom'];
-        $parent->Email = $validatedData['Email'];
-        $parent->Adress = $validatedData['Adress'];
-        $parent->tél = $validatedData['tél'];
-        $parent->Identifiant = $validatedData['Identifiant'];
+        $parent = ParentM::find($id);
+        $parent->Civilité =  $request->get('Civilité');
+        $parent->Nom =  $request->get('Nom');
+        $parent->Prénom =  $request->get('Prénom');
+        $parent->Email =  $request->get('Email');
+        $parent->Adress = $request->get('Adress');
+        $parent->tél =  $request->get('tél');
+        $parent->Identifiant =  $request->get('Identifiant');
         $parent->save();
 
         return redirect()->route('parents.index')->with('success', 'Le parent a été modifié avec succès.');
     }
 
-    public function destroy($id)
-    {
-        $parent = ParentM::findOrFail($id);
-        $parent->delete();
+       public function destroy($id)
+{
+    $parent = ParentM::find($id);
 
-        return redirect()->route('parents.index')->with('success', 'Le parent a été supprimé avec succès.');
+    if (!$parent) {
+        return response()->json(['message' => 'Le parent n\'existe pas'], 404);
     }
+
+    $parent->delete();
+    return response()->json(['message' => 'Le parent a été supprimé avec succès']);
+}
+
 }
